% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinetr.R
\docType{package}
\name{epinetr}
\alias{epinetr}
\title{epinetr}
\description{
epinetr is a package intended to aid in the investigation of the
contribution of epistatic networks to complex traits,
}
\details{
This package provides a range of functions for running
forward-time simulation using epistatic networks, including
visualisation tools.

For a complete list of functions, use library(help = "epinetr").
}
\author{
Dion Detterer, Paul Kwan, Cedric Gondro
}
