% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geno100snp}
\alias{geno100snp}
\title{Example genotype matrix.}
\format{
\code{geno100snp} is a matrix with 500 rows and 200
columns, used in examples when constructing a population from
genotypes using 500 individuals and 100 SNPs.
}
\usage{
geno100snp
}
\description{
An example genotype matrix for 100 SNPs across 500 individuals.
}
\seealso{
\code{\link{Population}}, \code{\link{map100snp}}
}
\author{
Dion Detterer, Paul Kwan, Cedric Gondro
}
\keyword{datasets}
