% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map100snp}
\alias{map100snp}
\title{Example map.}
\format{
\code{map100snp} is a \code{data.frame} with 100 rows and
3 variables:
\describe{
  \item{V1}{SNP ID}
  \item{V2}{chromosome ID}
  \item{v3}{position on chromosome, in base pairs}
}
}
\usage{
map100snp
}
\description{
An example map \code{data.frame} for 100 SNPs across 22
chromosomes.
}
\seealso{
\code{\link{Population}}, \code{\link{geno100snp}}
}
\author{
Dion Detterer, Paul Kwan, Cedric Gondro
}
\keyword{datasets}
