% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rincmat100snp}
\alias{rincmat100snp}
\title{Example incidence matrix.}
\format{
\code{rincmat100snp} is a matrix with 20 rows and 19
columns, used in examples when constructing an epistatic network
using a user-supplied incidence matrix.
}
\usage{
rincmat100snp
}
\description{
An example incidence matrix for 19 pairwise interactions across
20 QTLs.
}
\seealso{
\code{\link{attachEpiNet}}
}
\author{
Dion Detterer, Paul Kwan, Cedric Gondro
}
\keyword{datasets}
